Extracts workflow related details and moved them to a sep. detail file
