+++ webofneeds/won-owner-webapp/src/main/webapp/config/details/files.js
@@ -167,103 +167,3 @@ export const images = {
     return undefined;
   },
 };
-export const bpmnWorkflow = {
-  identifier: "bpmnWorkflow",
-  label: "BPMN",
-  icon: "#ico36_detail_workflow",
-  placeholder: "",
-  //accepts: "application/octet-stream",
-  accepts: "",
-  component: "won-workflow-picker",
-  viewerComponent: "won-workflow-viewer",
-  messageEnabled: true,
-  parseToRDF: function({ value, identifier, contentUri }) {
-    if (value && value.name && value.data) {
-      //do not check for value.type might not be present on some systems
-      let workflow = {
-        "@id":
-          contentUri && identifier
-            ? contentUri + "/" + identifier + "/" + generateIdString(10)
-            : undefined,
-        "@type": "s:FileObject",
-        "s:name": value.name,
-        "s:type": value.type,
-        "s:data": value.data,
-      };
-
-      return { "won:hasBpmnWorkflow": workflow };
-    }
-    return { "won:hasBpmnWorkflow": undefined };
-  },
-  parseFromRDF: function(jsonLDImm) {
-    const wflw = jsonLDImm && jsonLDImm.get("won:hasBpmnWorkflow");
-
-    let workflow = {
-      name: get(wflw, "s:name"),
-      type: get(wflw, "s:type"),
-      data: get(wflw, "s:data"),
-    };
-    if (workflow.name && workflow.data) {
-      //do not check for value.type might not be present on some systems
-      return Immutable.fromJS(workflow);
-    }
-
-    return undefined;
-  },
-  generateHumanReadable: function({ value, includeLabel }) {
-    if (value && value.name) {
-      return includeLabel ? this.label + ": " + value.name : value.name;
-    }
-    return undefined;
-  },
-};
-export const petrinetWorkflow = {
-  identifier: "petrinetWorkflow",
-  label: "Petrinet",
-  icon: "#ico36_detail_workflow",
-  placeholder: "",
-  //accepts: "application/octet-stream",
-  accepts: "",
-  component: "won-petrinet-picker",
-  viewerComponent: "won-petrinet-viewer",
-  messageEnabled: true,
-  parseToRDF: function({ value, identifier, contentUri }) {
-    if (value && value.name && value.data) {
-      //do not check for value.type might not be present on some systems
-      let workflow = {
-        "@id":
-          contentUri && identifier
-            ? contentUri + "/" + identifier + "/" + generateIdString(10)
-            : undefined,
-        "@type": "s:FileObject",
-        "s:name": value.name,
-        "s:type": value.type,
-        "proc:hasInlinePetriNetDefinition": value.data,
-      };
-
-      return { "won:hasPetrinet": workflow };
-    }
-    return { "won:hasPetrinet": undefined };
-  },
-  parseFromRDF: function(jsonLDImm) {
-    const wflw = jsonLDImm && jsonLDImm.get("won:hasPetrinet");
-
-    let workflow = {
-      name: get(wflw, "s:name"),
-      type: get(wflw, "s:type"),
-      data: get(wflw, "wf:hasInlinePetriNetDefinition"),
-    };
-    if (workflow.name && workflow.data) {
-      //do not check for value.type might not be present on some systems
-      return Immutable.fromJS(workflow);
-    }
-
-    return undefined;
-  },
-  generateHumanReadable: function({ value, includeLabel }) {
-    if (value && value.name) {
-      return includeLabel ? this.label + ": " + value.name : value.name;
-    }
-    return undefined;
-  },
-};
